Move singletonColumnVector
