+++ kernel/kernel-api/src/main/java/io/delta/kernel/internal/util/InternalUtils.java
@@ -21,8 +21,11 @@ import java.time.LocalDate;
 import java.time.temporal.ChronoUnit;
 import java.util.Optional;
 
+import io.delta.kernel.data.ColumnVector;
 import io.delta.kernel.data.FileDataReadResult;
 import io.delta.kernel.data.Row;
+import io.delta.kernel.types.DataType;
+import io.delta.kernel.types.StringType;
 import io.delta.kernel.utils.CloseableIterator;
 
 public class InternalUtils {
@@ -124,5 +127,42 @@ public class InternalUtils {
     public static int daysSinceEpoch(Date date) {
         LocalDate localDate = date.toLocalDate();
         return (int) ChronoUnit.DAYS.between(EPOCH, localDate);
+    }
+
+    /**
+     * Utility method to create a singleton string {@link ColumnVector}
+     *
+     * @param value the string element to create the vector with
+     * @return A {@link ColumnVector} with a single element {@code value}
+     */
+    public static ColumnVector singletonStringColumnVector(String value) {
+        return new ColumnVector() {
+            @Override
+            public DataType getDataType() {
+                return StringType.STRING;
+            }
+
+            @Override
+            public int getSize() {
+                return 1;
+            }
+
+            @Override
+            public void close() {
+            }
+
+            @Override
+            public boolean isNullAt(int rowId) {
+                return value == null;
+            }
+
+            @Override
+            public String getString(int rowId) {
+                if (rowId != 0) {
+                    throw new IllegalArgumentException("Invalid row id: " + rowId);
+                }
+                return value;
+            }
+        };
     }
 }
